backend/backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
python-dotenv